requirements.txt
streamlit
google-generativeai
python-dotenv
pdf2image
pandas
Pillow
openai
openpyxl
# For hybrid mode (template + LLM):
# Install invoice2data from the local repo: pip install -e invoice2data_repo
# Or install from PyPI: pip install invoice2data
